feat(events-util): Relocating getEventTimestamp util
