+++ static/app/components/events/interfaces/performance/spanEvidenceTraceView.tsx
@@ -1,10 +1,8 @@
 import {lazy, Suspense, useMemo} from 'react';
 import styled from '@emotion/styled';
 
-import {
-  getTimestampFromEvent,
-  TRACE_WATERFALL_PREFERENCES_KEY,
-} from 'sentry/components/events/interfaces/performance/utils';
+import {TRACE_WATERFALL_PREFERENCES_KEY} from 'sentry/components/events/interfaces/performance/utils';
+import {getEventTimestampInSeconds} from 'sentry/components/events/interfaces/utils';
 import type {Event} from 'sentry/types/event';
 import type {Organization} from 'sentry/types/organization';
 import {useIssuesTraceTree} from 'sentry/views/performance/newTraceDetails/traceApi/useIssuesTraceTree';
@@ -77,7 +75,7 @@ function SpanEvidenceTraceViewImpl({
   organization,
   traceId,
 }: SpanEvidenceTraceViewProps) {
-  const timestamp = getTimestampFromEvent(event);
+  const timestamp = getEventTimestampInSeconds(event);
 
   const trace = useTrace({
     timestamp,
