+++ static/app/components/quickTrace/utils.tsx
@@ -1,5 +1,6 @@
 import type {Location, LocationDescriptor} from 'history';
 
+import {getEventTimestampInSeconds} from 'sentry/components/events/interfaces/utils';
 import {normalizeDateTimeParams} from 'sentry/components/organizations/pageFilters/parse';
 import {ALL_ACCESS_PROJECTS} from 'sentry/constants/pageFilters';
 import type {Event} from 'sentry/types/event';
@@ -82,28 +83,6 @@ export function generateSingleErrorTarget(
   }
 }
 
-const timestampsFieldCandidates = [
-  'dateCreated',
-  'startTimestamp',
-  'timestamp',
-  'endTimestamp',
-];
-
-export function getEventTimestamp(event: Event): string | number | undefined {
-  for (const key of timestampsFieldCandidates) {
-    if (
-      key in event &&
-      // @ts-expect-error TS(7053): Element implicitly has an 'any' type because expre... Remove this comment to see the full error message
-      (typeof event[key] === 'string' || typeof event[key] === 'number')
-    ) {
-      // @ts-expect-error TS(7053): Element implicitly has an 'any' type because expre... Remove this comment to see the full error message
-      return event[key];
-    }
-  }
-
-  return undefined;
-}
-
 export function generateTraceTarget(
   event: Event,
   organization: Organization,
@@ -120,7 +99,7 @@ export function generateTraceTarget(
       organization,
       traceSlug: traceId,
       dateSelection,
-      timestamp: getEventTimestamp(event),
+      timestamp: getEventTimestampInSeconds(event),
       eventId: event.eventID,
       location,
       source,
